requirements.txt
asttokens==2.0.5
attrs==22.1.0
backcall==0.2.0
black==22.6.0
click==8.1.3
decorator==5.1.1
executing==0.9.1
jedi==0.18.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
pygame==2.1.2
Pygments==2.12.0
six==1.16.0
stack-data==0.3.0
tomli==2.0.1
traitlets==5.3.0
typing_extensions==4.3.0
wcwidth==0.2.5
